splitted analyzer test by statement
with common setup stuff in superclass
